Fix %d/unsigned long printf mismatch found by clang
In order to enable clang-cl's printf format string mismatch checking we
need to fix existing warnings, even benign ones like this Windows-
specific %d/unsigned long mismatch (benign because unsigned long is
always int-sized on Windows).

Normally the best way to fix these mismatches is to modify the format
string, but the process ID type is, depending on platform, either
DWORD, mx_koid_t, or pid_t, with the possibility that one or more of
these may vary depending on 32-bit versus 64-bit. So, a cast to int
is the pragmatic fix, just this once.

R=dgozman@chromium.org
BUG=751171

+++ content/browser/devtools/service_worker_devtools_agent_host.cc
@@ -68,8 +68,8 @@ std::string ServiceWorkerDevToolsAgentHost::GetType() {
 
 std::string ServiceWorkerDevToolsAgentHost::GetTitle() {
   if (RenderProcessHost* host = RenderProcessHost::FromID(worker_id().first)) {
-    return base::StringPrintf("Worker pid:%d",
-                              base::GetProcId(host->GetHandle()));
+    return base::StringPrintf(
+        "Worker pid:%d", static_cast<int>(base::GetProcId(host->GetHandle())));
   }
   return "";
 }
